+++ database/migrations/2021_12_14_081337_create_comments_table.php
@@ -16,7 +16,7 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->increments('id');
             $table->string('comment');
-            $table->double('rating');
+            $table->float('rating');
             $table->integer('restaurant_id')->unsigned();
             $table->integer('user_id')->unsigned();
             $table->timestamps();
